Moved all startup.py things to OpenPyCommands.py
